requirements.txt
boto3==1.4.4
botocore==1.5.80
dj-database-url==0.3.0
dj-static==0.0.6
django-extensions==1.9.6
Django==1.11
Pillow==3.2.0
psycopg2==2.6.1
static3==0.7.0
gunicorn==19.6.0
django-tastypie==0.14.2
python-dateutil==2.6.0
python-mimeparse==1.6.0
defusedxml==0.5.0
lxml==4.0.0
django-phonenumber-field==1.3.0
records==0.5.2